refactor: move priority tab to ui components
